Root: implement backend-agnostic methods
All of the methods except Root::resolve can be implemented in a
backend-agnostic manner. This does mean that O_CREAT (with a trailing
symlink) won't work, but we can always move create_file() to be a
required method.

In addition, $crate::Handle was switched to be entirely
backend-independent since it will always be implemented in the smae
manner ("/proc/self/fd/$n" re-open, or O_EMPTYPATH if that's ever merged
upstream).

Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>