ffi: switch to error struct to return errno
Since I imagine quite a few users will really care about the errno
value, we should expose it through C FFI. This also has the benefit of
removing the ugly need for allocation functions. Yes, it limits the size
of error strings but that's not really the end of the world.

Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>