pathrs: harden /proc usage
Instead of assuming that /proc is actually procfs, we open it with
O_PATH on first usage and do the relevant fstatfs(2) check -- and then
re-use that handle for the rest of our program lifetime. Yes, most other
people don't do this (since it's quite overkill) but it's simple enough
to implement with lazy_static! and there's no real downside.

In order to get this to work, we switch the API away from using
OpenOptions -- because it's not currently possible to get the set of
flag bits it represents[1].

[1]: https://github.com/rust-lang/rust/issues/62972

Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>