lib: fix pretty-printing of syscall fd arguments
Because of Rust's drop semantics, returning an error that references
the RawFd associated with a file results in the pretty-printing always
failing and giving "<unknown>". While we could instead include the
actual File and thus depend on normal drop semantics working nicely,
this would cause problems with functions that only have &File -- and it
would be significantly uglier. So instead we just save the
pretty-printed string when the error is generated.

Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>