ffi: associate errors with pathrs objects
Having a global error value which is cleared-on-read is not the best
idea in the world because it makes it more tricky to deal with the error
value properly (effectively the same problem as errno). So instead,
associate the error with the object itself to avoid this problem -- this
is a fairly classic C interface.

Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>