utils: further harden /proc handle
Rather than just checking the filesystem type, also ensure that it's the
root of a procfs mount. Luckily, procfs does actually guarantee a stable
inode value (which can't be faked by an attacker).

There are still potential mount-over attacks, but I think we can solve
them in the future by doing a manual path walk (or RESOLVE_NO_XDEV). But
we'll have to write helpers for all procfs operations. Also we need
handle readlinkat(2) specially -- because you can't do a readlink of an
O_PATH to a symlink.

Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>