ffi: re-work error handling for pathrs_open
It's necessary for us to have some way of returning errors from
pathrs_open(), but given the (fairly standard) object-based error setup
with libpathrs there isn't an object to tie a pathrs_open() error to.

So the (fairly ugly) solution is to return a "fake" pathrs_root_t in
case of an error, and use it to store the error. Attempting to use the
dummy object anywhere else will always result in an error. This does
make the API a little bit fruity:

    pathrs_root_t *root = pathrs_open("/foo");
    error = pathrs_error(PATHRS_ROOT, root);
    if (error)
        goto err;

    pathrs_handle_t *handle = pathrs_resolve(root, "/bar/baz");
    if (!handle) {
        error = pathrs_error(PATHRS_ROOT, root);
        goto err;
    }

    int ret = pathrs_creat(root, "/bar/buzz");
    if (ret < 0) {
        error = pathrs_error(PATHRS_ROOT, root);
        goto err;
    }

But it's not clear if there is a better alternative (a thread-local
variable for the error would appear to be the most reasonable, but will
immediately run into problems with M-to-N systems like Go or async-aware
languages).

However, the one benefit of this approach is that users /could/ do
something like this (even though it's still a bit odd):

    pathrs_root_t *root = pathrs_open("/foo");
    error = pathrs_error(PATHRS_ROOT, root);
    if (error)
        goto err;

    pathrs_handle_t *handle = pathrs_resolve(root, "/bar/baz");
    error = pathrs_error(PATHRS_ROOT, root);
    if (error)
        goto err;

    int ret = pathrs_creat(root, "/bar/buzz");
    error = pathrs_error(PATHRS_ROOT, root);
    if (error)
        goto err;

Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>