ffi: make pathrs_configure() forward and backward compatible
First of all, I just want to apologise to the reader for some of the
ugly hacks needed to work around cbindgen and Python's cffi. Turns out
__attribute__((packed)) isn't as widely supported as you might think.

The basic idea is to copy openat2(2)'s design for libpathrs. The only
API which takes a possibly-extensible struct is pathrs_configure() and
so we only need to have this (slight) API wart for that.

In future, we might actually implement a 'static inline' wrapper to
avoid getting users to do sizeof (though this would need to be done
before we actually cement the API of libpathrs):

  static inline pathrs_error_t *
  pathrs_configure(pathrs_type_t ptr_type, void *ptr,
                   void *old_cfg_ptr, const void *new_cfg_ptr)
  {
    size_t cfg_size = 0;
    switch (ptr_type) {
    case PATHRS_NONE:
      cfg_size = sizeof(pathrs_config_global_t);
      break;
    case PATHRS_ROOT:
      cfg_size = sizeof(pathrs_config_root_t);
      break;
    }

    return __pathrs_configure(ptr_type, ptr,
                              old_cfg_ptr, new_cfg_ptr, cfg_size);
  }

Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>