contrib: bindings/python: complete rework
Rather than relying on all of the runtime-based (and very crash-happy)
bindings using the ABI mode of cffi, use the API mode. This does require
a pre-build step but has the major benefits of actually linking to
libpathrs.so and being much more resilient to runtime crashes due to
type (or other) errors.

Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>