root: drop stored path in Root
The original purpose of this was to avoid cases where the root was moved
and the resulting lookup could be deemed unsound. However, this check
was not sufficient to protect against this[1] and also made it harder
to use libpathrs[2].

We do include a basic double-check for the root path during userspace
resolution, but users should be aware that this is just to make attacks
a bit harder -- users should not permit attackers to control the Root
directory.

[1]: The attacker could just move the root back in between the lookups
     and checks, thus potentially winning the race. We don't have a
     seqlock-like equivalent in userspace so all we can detect if the
     value is *different* (not if it has been *modified* since the last
     read).

[2]: Passing the fully-resolved path may seem trivial but shells will
     happily give you paths containing symlinks if you use the shell's
     builtin "pwd" command (this is the case for a bunch of historical
     reasons out of scope here).

Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>