root: make resolver configurable with flags
In order to allow certain extensions, it'd be a good idea to allow users
to specify any additional flags they'd like to be passed to openat2(2)
(or emulated with the userspace resolver).

This requires another cbindgen patch to support repr(u64)[1]. Luckily,
they just released cbindgen 0.12.0 with our repr(align(...)) patches.

This is also a pre-requisite of some of the procfs hardening work.

[1]: https://github.com/eqrion/cbindgen/pull/441

Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>