{Root,Handle}: add serialise and deserialise methods
By adding try_clone, into_file, and from_file_unchecked it is now
possible for most programs to successfully pass around Roots and
Handles (even with SCM_RIGHTS). Once we add a C API for this, we can
consider that feature complete.

The reason for avoiding Into<File> and From<File> is that they are not
semantically the correct way to represent these conversions (especially
since from_file_unchecked has certain safety guarantees the caller needs
to make). And Clone is not usable because we need to return an error if
the File dup fails.

Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>