contrib: bindings: add go binding
This binding is mainly inspired by python building.
The public interface is built on the shoulders of `Root` and `Handle`
structures, in a similar way as the original library does.
These ones are straightforward wrappers of the C api to provide
a safe and idiomatic interface.

As for error handling it upholds the idea to copy all the data which is
provided by ffi interface including backtrace.
It copies it to an internal structure to not being bound
by handling errors on place. Since golang does not has a
standart way to manage stackraces it can be get by casting
error to `pathrs.Error` type and call `Backtrace` method.

This binding does not provide configuration interface of pathrs.

Signed-off-by: Maxim Zhiburt <zhiburt@gmail.com>