contrib: bindings/go: rework IntoFile interface
It was a mistake on my part to suggest cloning the handle when doing
execution, because libpathrs is already designed to handle a
pathrs_free() on an object which has been pathrs_into_fd()'d.

Additionally, the interfaces for Handle and Root were not the same
(making things more confusing). The naming was also a bit odd, so I've
settled on IntoRaw and FromRaw because it better conveys that the handle
you get is a "raw" handle and is different from the one you get from
handle.Open().

Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>