capi: make pathrs_from_fd() clone the file
Some GC'd language runtimes might not be able to (trivially) uphold the
requirement of Rust owning the lifetime of a file handle which was
created in the language. Not to mention that buggy programs could also
cause plenty of havoc.

Thus it is much simpler to manage our own handle.

Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>