capi: *: don't take &mut from C callers
Given that our library is intended to be used from multi-threaded
programs, we shouldn't lie to Rust by passing &muts. pathrs_free() must
be &mut, so we will need to figure out a way to provide the
multi-threading guarantee for pathrs_free (I hope it doesn't involve
locking in the caller).

Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>