capi: redesign CPointer locking and error handling
This is not a user-visible change, but this entirely reworks (and makes
much cleaner) the way that locking is handled for CPointer objects. The
original locking looked something like this:

  Arc / RwLock / GC [1]
  -- FFI Boundary --
    Arc
      Mutex
        inner:      Option<T>
        last_error: HashMap<ThreadId, Option<Error>>

The immediate problem with this design is that all operations are
serialised through the same lock, even though most operations on the
inner object can be done with a read-only reference. It also was a
little unwieldy to use, and Arc<Mutex<...>> is not necessary in this
context. This patch updates the locking to look more like:

  Arc / RwLock / GC [1]
  -- FFI Boundary --
    RwLock
      inner: Option<T>
    Mutex
      last_error: HashMap<ThreadId, Option<Error>>

By using an RwLock on the inner object, we can now take advantage of the
fact that most operations on our objects are multi-thread safe. The lock
for last_error was split because every operation needs write access to
the HashMap (to clear or set the thread-specific error), and so in order
to get the RwLock advantages we need to split it.

It should be noted that Mutex<HashMap<...>> is fairly non-ideal in most
cases (serialising all accesses through a single lock kills performance)
and probably should be switched to a more concurrent HashMap-like
structure (such as chashmap[2]). Projects such as evmap[3] won't help
much because our workload is very write-heavy -- *but* each thread only
writes to a single slot in the HashMap, making granular-locked
structures much more appealing.

[1]: The top-level layer is necessary because the C caller needs to
     guarantee &mut-equivalent exclusion for pathrs_free() otherwise
     double-frees will occur.
[2]: https://crates.io/crates/chashmap
[3]: https://crates.io/crates/evmap

Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>