capi: switch to fd-based API
Using leaked Box<T>s with quite complicated semantics for handling
errors and freeing doesn't make much sense when we are mostly operating
on file descriptors. Switching to a kernel-like API where fds are
returned as numerical values and errors are negative values is a nicer
interface for C programs (and file descriptors are easier to write
bindings for -- especially for Go since *os.File is quite magical).

However, a key question is how to handle errors. Having per-thread
errors (a-la errno) makes it far annoying to write bindings for
green-thread-based languages like Go. So, return a random negative
number that is used as a reference in a global mapping of error numbers
to underlying errors. The error is removed from the map when the C user
gets the error information. This could cause issues if a C user never
checked their error messages, but that can easily be avoided by not
doing that.

There are some possible lock contention issues and ideally we would use
a hashmap with internal key-based locking (which would eliminate lock
contention entirely) but since this is only for the error path we can
look at optimising this later.

In short, the locking has gone from

  Arc / RwLock / GC [1]
  -- FFI Boundary --
    RwLock
      inner: Option<T>
    Mutex
      last_error: HashMap<ThreadId, Option<Error>>

to

  fd: no locking in libpathrs
  err: global Mutex<HashMap<c_int, Error>>

We also have to drop the configuration API because we no longer have
opaque objects that can can have config settings attached to them. We
will need to re-add this in the future (most likely in the form of a
pathrs_..._c family of functions that take a config struct as an
argument) but for now this should be okay.

Suggested-by: Christian Brauner <brauner@kernel.org>
Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>