go bindings: completely rewrite Go bindings to new API
The new API is so different that it requires a complete rewrite and
removes almost all of the cruft that was needed to avoid double-closes
and thread-migration with Go.

As this is a complete rewrite, the copyright line of the original author
of the bindings is removed (arguably the bindings were completely
changed with the last rewrite but in this case they were actually
completely rewritten).

Also, change the import path to the slightly nicer

  import "github.com/openSUSE/libpathrs/go-pathrs"

We don't have proper pkg-config definitions nor install scripts yet, so
the Go binding still needs to do some manual fiddling with CFLAGS and
LDFLAGS for now.

Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>