python bindings: update bindings to use new API
We can't use fdopen() for the new file descriptor API for two reasons:

 1. fdopen() doesn't work for directory file descriptors.
 2. io.IOBase file descriptors are always closed when garbage collected,
    but some users may want to leak the file in order to just use the
    file descriptor directly.

So we need to have our own WrappedFd that does GC-based closing by
default but lets you leak the file descriptor. When constructing a
WrappedFd we clone the underlying file descriptor if the source was an
io.IOBase to ensure the GC won't close the file from underneath us.

The constructors for Root and Handle are now a little more python-like
(you can pass any file-like thing and they will handle it properly, and
you can still pass a string path to Root to open a new root).

Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>