capi: remove backtraces from FFI API
Backtraces are nice for debugging, but including them in the C API
overcomplicates the API for very little benefit.

The original intention was to be able to splice our Rust backtraces into
backtraces for binding languages (like Python and Go). However, it turns
out this is not really possible for those langauges and it seems
unlikely that anybody will actually make use of the backtrace
information we provide.

This also removes a lot of the annoying code we had for converting the
C representation of backtraces into something the binding language cares
about.

In the worst case, users that need to access backtraces for debugging
can always write a Rust version of the reproducer. libpathrs is a small
enough library that this should be possible for almost all reproducers
(not that backtraces would be necessary in most cases anyway).

Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>