deps: update to snafu v0.8
Most of the changes are from the v0.7 update we missed and were
mechanically applied with the snafu migration tool.

This switch also requires moving to std::backtrace::Backtrace, which
required a few other changes (including removing the global setting for
whether we want to capture backtraces).

If we really wanted to keep the global setting knob, we could wrap
std::backtrace::Backtrace in our own type but I don't think the global
setting really made sense in the first place (and now that there are no
backtraces in the C FFI, the standard RUST_* environment variables can
still be used for Rust programs).

Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>