resolvers: opath: work around Path::components() normalisation
It turns out that while Rust's Path::components() isn't as bad as Go's
filepath.Clean(), it will normalise "." components.

This results in us incorrectly not returning an error when there is a
"." component after a non-directory. While this is a correctness issue,
the larger issue is that this behaviour deviates from openat2, and we
need to keep openat2 compatibility.

Unfortunately this requires implementing our own Iterator over &OsStr,
as there are no helpful functions for creating such a generator in the
stdlib.

Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>