resolve: opath: handle // components and trailing-slash correctly
Previously we would just use them as though they are regular components,
resuling in ENOENT because openat doesn't allow "" opens.

When we hit empty components, we need to treat them as though they are a
"." component. We could skip them for non-trailing components but for
trailing components it's critical to actually try to do the open so that
we get openat2-like errors for non-directory states.

In addition, in order to detect a single trailing / when doing lookups,
we need to rework RawComponents a little bit to use an Option so that we
can return "" as a final component if there was a single trailing "/".

All of these behaviours are modelled after openat2's returns, and our
tests make sure they match.

Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>