resolvers: opath: use Rc to avoid extra root copies
This removes a lot of un-needed work that came from the fact that Go
doesn't have an equivalent of Rc<File>. In Rust we can just create a
single copy of the root handle.

We can be sure that there will only be at most two references to root
(we don't store the Rc anywhere else) and so once we drop the root Rc we
can be sure that the current Rc will only have one reference.

Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>