procfs: use a restricted procfs resolver for lookups
This is loosely based on the opath resolver, with support for using
openat2 if possible. The new procfs_beneath resolver currently only
supports doing O_PATH (and O_DIRECTORY) lookups. This is enough for our
needs within libpathrs but when we expose ProcfsHandle as a public API
we will need to support more modes. Unfortunately, we can't just do a
/proc/self/fd/... re-open because the resolver is used within the
ProcfsHandle implementation to do reopening.

Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>