procfs: add pre-3.17 fallback for /proc/thread-self
This logic is taken from ProcThreadSelf in runc and filepath-securejoin.
Some old RHEL systems have pre-3.17 kernels which require this
workaround.

Unfortunately, runc has to do /proc/... operations after joining a PID
namespace but while still using the host /proc, which results in errors
if we just try /proc/$pid/task/$tid (the TIDs are from our PID namespace
but the TIDs used by the procfs are based on the PID namespace used when
it was mounted, which means that host PID namespace in most cases). In
this case, the only other option is to just use /proc/self and hope
nothing breaks. We could panic, but users would probably then want to
use an API that does this fallback anyway (and there's no real way to
conclusively verify that /proc/self is "okay" to use).

Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>