resolvers: opath: improve symlink-related errors
We want to return -ELOOP for our RESOLVE_NO_SYMLINKS and
RESOLVE_NO_MAGICLINKS emulation (to match openat2). In addition, the
checking of whether a symlink is "safe" to follow for was overly
aggressive -- /proc/self is fine to follow, it's the absolute links that
don't make sense to follow.

Some new tests are included to verify this parity.

Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>