python bindings: implement context interface
Users probably will want to use our handles with the "with" statement.
We can also use it internally.

Switching to "with" in the static_web.py example also exposed a bug in
the error path of WrappedFd.fdopen(). Looking at the error path, it
makes more sense to "unleak" the file descriptor rather than try to
close it.

Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>