opath resolver: just use plain Rc<File> rather than RcFile
This is a revert of 7ea5c47757bc49bd244dca31609f46f6605ceb02.

It seems very unlikely the overhead of Rc is going to be a problem in
practice, and the SymlinkStack we need for Root::mkdir_all requires us
to convert RcFile::Original to RcFile::Ref which eliminates the major
benefit of RcFile (this is only relevant for partial lookups but the
code is kind of ugly regardless).

Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>