resolvers: *: implement resolve_partial()
This is needed in order to implement Root::mkdir_all() efficiently.
Unfortunately, for the opath resolver we need to create a "symlink
stack" data structure in order to make sure that we have the same
behaviour as the openat2 resolver when dealing with dangling symlinks.

This logic is mostly taken from filepath-securejoin. However, we treat
all errors (including -ELOOP) as though they are any other kind of
lookup error and return a partial lookup result in that case. Due to
some downsides of the Go implementation of the symlink stack,
filepath-securejoin only does this for -ENOENT and -ENOTDIR.

Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>