tests: resolve: rework macros
The resolve_tests! macro is pretty gnarly, but we can improve it a bit.
This will be needed for when we add partial resolution support.

We can also remove the resolve_tests boolean by opening /proc/self/sched
(which is a no-op writable procfs file) and thus doing the re-open tests
for procfs as well.

We also don't need the dedicated resolver::*::resolve() tests. The
Root::resolve tests are functionally identical, and doing both in the
same macro just makes the macro even uglier.

Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>