root: add mkdir_all() implementation
This is primarily necessary because Go users expect to be able to use
os.MkdirAll (which is not safe to use with an active attacker). We could
try to implement this just for Go within the bindings code, but it turns
out that you need to have a fairly intricate resolve_partial()
implementation for this to be efficient (especially for the opath
resolver).

This implementation is based on filepath-securejoin's safe MkdirAll
implementation.

Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>