msrv: fix at Rust 1.65
This version was released in 2022, so it should be old enough to use for
packaging in regular distributions.

Ideally we would try to use Rust 1.63 (the version in Debian stable),
but the main thing limiting us is std::backtrace::Backtrace. We do plan
to rework errors in the future (to avoid a bunch of the issues with
snafu), so we might make be able to get to 1.63 in the future.

We can't really go older than 1.63 because rustix requires it (likely
because of the BorrowedFd io_safety API), which we plan to migrate to as
well.

Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>