errors: switch to thiserror
This is much nicer to use from a Rust-y perspective. The only downside
is that (for now) we don't have backtraces. The documentation is a
little hazy but it seems like std::backtrace only works on nightly
because you need some not-yet-stabilised features.

Regardless, we plan to move our MSRV to 1.63 (which doesn't have
std::backtrace). We might be able to add backtrace (from the crate) to
our Error type but that means we can't use #[error(transparent)] (and we
would have to export the backtrace crate types because of 1.63). So
let's just punt on this for now...

Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>