tests: use pub(in crate::tests) as much as possible
The whole module is #[cfg(test)]-ed out, but Rust 1.63 doesn't take into
consideration whether or not a "pub fn" is actually exposed outside of
the module so it's best to explicitly not export stuff we aren't using.

The only test helper we use outside of src/tests is create_basic_root().

Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>