root: mkdir_all: return explicit errors if perms contains set[ug]id bits
We would already return an error in this case (because they are ignored
by mkdirat(2) so they wouldn't be applied and so our expected-mode
checks would fail) but since this silent-ignoring behaviour is intended
behaviour by Linux we should provide a more helpful error.

Users that want to have the silently-ignore behaviour can write a small
wrapper that masks the bits themselves.

Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>