root: mkdir_all: correctly handle sgid parent directory
When S_ISGID is set on a directory, it affects the ownership and mode of
inodes created underneath the directory (the owner gid of all inodes is
set to the directory's owner, and S_ISGID is applied to any new
directories created). We need to handle this properly in order to avoid
errors from our mode/owner validation logic.

Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>