root: resolvers: hide backend configuration
Since we don't expose the ability to change backends in the C API (and
it's not clear how many people would actually care about it), just
remove the ability for users to change the backend dynamically for now.

In principle we will want to let this be configurable, but the internal
API of ResolverBackend are a little ugly to give users at the moment.

Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>