python bindings: make cffi builds more friendly for distros
Some distributions (such as openSUS) would like to make the python
bindings a subpackage of the main libpathrs package, which means they
need to build the bindings and the library at the same time.

This is possible with setuptools, but being able to create wheels would
be ideal. The trick is to make it so that if you pass the magic
environment variable PATHRS_SRC_ROOT, we can use that as the source dir
for compiling against even when inside a venv for "python -m build".

This has no impact on the wheels we will make for PyPI.

Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>