python bindings: fix builds for pre-pyproject setuptools
openSUSE Leap only has setuptools 44, which predates support for
pyproject. This is fine, we just need to add the relevant metadata to
the setuptools invocation. To reduce the amount of places we duplicate
this information, just load pyproject.toml and read the values out (this
does add a dependency on toml for pre-3.11 Python versions but that's
fine).

Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>