procfs: create temporary handle for non-subset=pid subpaths
We want to set subset=pid,hidepid=ptraceable for the global procfs
handle (to restrict the scope if the handle was leaked into a
container). Unfortunately we also need to access stuff outside of
/proc/$pid such as sysctls (/proc/sys/*), and we plan to allow users to
access stuff in the root of /proc eventually.

So, in order to balance these two requirements, we can just create a
temporary procfs handle whenever we need to access stuff outside of the
root (which should be far more rare than accessing things inside
/proc/$pid) and use it as a fallback if the global procfs handle return
ENOENT. This will increase the overhead of accessing stuff outside
/proc/$pid, but the security benefits should be well worth it.

Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>