capi: procfs: use open-enum to detect invalid enum values
Even with #[non_exhaustive], it is Undefined Behaviour™ to instantiate
a Rust enum with an invalid value. Since C callers can pass any integer
they like to pathrs_proc_*(), this could cause serious confusion.

open-enum makes this fairly easy to check and has very few dependencies,
so we can just switch to that. Luckily cbindgen doesn't seem to care
about the #[open_enum] proc macro and still outputs a reasonable-looking
header file (I suspect this wouldn't be the case if we open-coded it).
Unfortunately, due to our MSRV we have to use a fairly outdated version.

Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>