python bindings: fix bugs noticed by mypy
Most of these are because of function name changes we made but forgot to
update all of the callers. There are also a few typing errors -- the
only one we don't fix at the moment is Root.creat() because it turns out
all our APIs (including Rust and Go) return a Handle when actually you
want to return a regular file for this case. This will be fixed in a
separate commit.

Fixes: c7728ed94781 ("python bindings: switch to setuptools")
Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>