python bindings: make type annotations for CString and CBuffer less magical
In order to avoid having magical types that only exist during
type-checking, define CString and CBytes in the main _pathrs module and
import them into the _libpathrs_cffi type stubs.

Unfortunately, the straight-forward way of doing this (using a TypeAlias
of _libpathrs_cffi.ffi.CData) doesn't work for some reason so we need to
do a little bit of magic when TYPE_CHECKING to make mypy happy. But at
least this is a little bit less magical than the previous version (where
some variables only existed when type-checking).

Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>