python bindings: make pydoc happy with package metadata
In order for help(pathrs) to correctly list all of the re-exported code
from pathrs._pathrs we need to add the names to __all__ in the main
module. Rather than duplicating the __all__ declaration we can just
re-use the one defined in pathrs._pathrs.

In addition, to avoid having __version__ drift away from the other
version definitions, we can use importlib.metadata.version() to use the
dist-info/ version (defined in pyproject.toml) so that pydoc just
reflects the actual dist-info/ version information.

Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>