git: hide python reformatting commit from git-blame
commit c2924e9e77d2 ("python bindings: reformat with black") messes with
the history, so we should hide it from git-blame as best we can. The
filename ".git-blame-ignore-revs" appears to have become a de-facto
standard and tools like GitHub have native support for it.

In order to have this work on your local machine, you need to configure
your .gitconfig to use this file:

  % git config blame.ignoreRevsFile .git-blame-ignore-revs

This feature was added in Git 2.23.

Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>