python bindings: add Root.creat_raw and sanify Root.creat API
Users probably want to be able to do Root.creat and get a WrappedFd, so
add a method for that.

In addition, the argument order for Root.creat didn't match the
openat(O_CREAT) argument order, which could cause some confusion so
rotate the arguments to match (this means we have to give filemode a
default value -- 0o644 makes the most sense as a default).

Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>