cargo: commit lockfile
Rust has changed their recommendation on whether libraries should commit
lockfiles. It seems the answer is now that we should err on the side of
committing it unless there is a strong reason not to[1]. Given that we
also package libpathrs for distros, I suspect most would prefer a
lockfile to improve reproducibility a bit (even if they can track
dependencies themselves).

CI will no longer test the latest version of every dependency, but we
have dependabot so that just means we'll see test results in the
(slightly spammy) PRs provided by dependabot.

[1]: https://blog.rust-lang.org/2023/08/29/committing-lockfiles.html

Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>