rust 2018: migrate to 2018-style modules
We can now move mod.rs to be <module_name>.rs which is easier to keep
track of. However, in some cases it seemed nicer to remove the mod.rs
file entirely and inline it in the upper module (this is especially nice
for mod.rs files that are completely trivial and just define the list of
submodules -- having to edit separate files is a bit of a waste of
time).

Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>