Root: do not error out with EEXIST for racing Root::mkdir_all()s
If two programs are doing Root::mkdir_all, the previous logic would
return an error if a directory already existed once we got into the
"mkdir" portion of the creation.

Since we already have to accept that an attacker can swap the inode with
a different directory, returning -EEXIST from mkdirat(2) just causes
spurious errors. All we care about is that we open a directory.

Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>