openat2: always return hard SafetyViolation for partial lookups
The opath resolver will return SafetyViolation if it notices a safety
violation, but because openat2 just gets -EXDEV (and resolve_partial is
implemented in terms of resolve) we ended up returning the safety
violation as a partial resolution.

In theory, returning a partial resolution is fine (since users of
partial-resolutions need to be race-safe anyway) but it is preferable to
avoid an attacker being able to trick libpathrs into continuing to
execute after we've detected an attack. This also unifies our behaviour.

In addition, remove the unused code at the end of the partial resolver
(coverage analysis shows that this code is never executed and we can be
fairly sure of that).

Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>