utils: remove_all: handle ENOENT more gracefully
It is possible for multiple threads to run Root::remove_all at the same
time, and the preferred behaviour should be that if the directory is
removed we treat that as being a success even if it wasn't removed by us.

This also means that trying to remove a non-existent directory should
also succeed -- this matches the os.RemoveAll behaviour we want to
mirror (as well as 'rm -rf'). This doesn't match the behaviour of
std::fs::remove_dir_all, but that is a bit of an outlier.

Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>