Root::resolve_parent: unify simple trailing-slash handling
Most of the users of Root::resolve_parent have very simple
trailing-slash handling requirements, so it makes sense to just move the
trailing slash stripping to Root::resolve_parent and update callers to
handle it.

The one thing to note is that Root::remove_all will now completely
ignore trailing slashes -- the reason for this is simply that ignoring
them is much simpler than the alternative, and is arguably more
consistent (as Root::remove_all is meant to mirror os.RemoveAll and "rm
-rf").

Root::rename needs more complicated handling, and so will be fixed in a
future patch in this series.

Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>