Root::rename: match trailing-slash handling to renameat2(2)
As opposed to other filesystem operations, renameat2(2) has slightly
peculiar trailing-slash semantics, and so replicating them requires a
bit of extra work.

One obvious thing to note is that we are doing fstatat(2) checks of the
paths, but an attacker could change them after the fact. This is
completely fine, since the error checking is mainly meant to be a
quality of life improvement over just ignoring trailing slashes
entirely. If an attacker changes the inode to no longer be a directory,
renameat2(2) itself will fail with an appropriate error.

Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>